simplementation of rollup

+++ statis/compile.js
@@ -1,6 +1,8 @@
 const fs = require('fs-extra')
 const plugins = require('./plugins.js')
+const runRollup = require('./run-rollup.js')
 const path = require('path')
+const state = require('./state.js')
 
 const compileTypes = {
   import: addImport,
@@ -9,11 +11,11 @@ const compileTypes = {
   style: addStyle
 }
 
-module.exports = function ({ fileContent, fileName }) {
-  return compiler(fileContent)
+module.exports = function ({ fileContent, fileObj, fileName }) {
+  return compiler(fileContent, fileObj, fileName)
 }
 
-function compiler (fileContent, fileName) {
+function compiler (fileContent, fileObj, fileName) {
   // console.log('before content:', fileContent)
   const varRegex = /<<([^<>]+)=([^<>]+)>>/igm
   fileContent = fileContent.replace(varRegex, (match, p1, p2) => {
@@ -21,7 +23,7 @@ function compiler (fileContent, fileName) {
     const command = { func: p1.trim(), val: p2.trim() }
     if (Object.keys(compileTypes).includes(command.func)) {
       console.log('got f', command)
-      return compileTypes[command.func](command.val, fileName) || match
+      return compileTypes[command.func](command.val, fileObj, fileName) || match
     } else {
       return match
     }
@@ -33,24 +35,30 @@ function addPath (filePath) {
   return filePath
 }
 
-function addScript (filePath) {
+function addScript (filePath, fileObj) {
   // Need to runPlugins()
-  const newPath = path.join('js', path.parse(filePath).base)
+  const newPath = path.join('js', path.parse(filePath).name + '.js')
   const scriptContent = fs.readFileSync(filePath, 'utf-8')
   const publicPath = path.join('public', newPath)
-  fs.ensureDirSync(path.join('public', 'js'))
-  fs.writeFileSync(publicPath, scriptContent)
+  if (!alreadyCompiled(filePath)) {
+    runRollup(scriptContent, fileObj, newPath).then(result => {
+      fs.ensureDirSync(path.join('public', 'js'))
+      fs.writeFileSync(publicPath, result.fileContent)
+    })
+  }
   return '<script src="' + newPath + '"></script>'
 }
 
 function addStyle (filePath) {
   // Need to runPlugins()
-  const newPath = path.join('css', path.parse(filePath).base)
+  const newPath = path.join('css', path.parse(filePath).name + '.css')
   const scriptContent = fs.readFileSync(filePath, 'utf-8')
   console.log('CONTEN', scriptContent)
   const publicPath = path.join('public', newPath)
-  fs.ensureDirSync(path.join('public', 'css'))
-  fs.writeFileSync(publicPath, scriptContent)
+  if (!alreadyCompiled(filePath)) {
+    fs.ensureDirSync(path.join('public', 'css'))
+    fs.writeFileSync(publicPath, scriptContent)
+  }
   return '<link rel="stylesheet" href="' + newPath + '">'
 }
 
@@ -62,9 +70,10 @@ function addImport (filePath) {
       console.log('exist')
       const fileObj = path.parse(filePath)
       const fileContent = fs.readFileSync(filePath, 'utf-8')
-      const newFile = plugins(fileContent, fileObj, filePath)
+      const newFile = runRollup(fileContent, fileObj, filePath) // need to use then
+      console.log('sfagfd', newFile.fileContent)
       newFile.fileContent = compiler(newFile.fileContent)
-      return newFile.fileContent
+      return newFile.fileContent // and return promise
     } else {
       return false
     }
@@ -74,3 +83,6 @@ function addImport (filePath) {
   }
 }
 
+function alreadyCompiled (filePath) {
+  return state.filesBuilt.includes(filePath)
+}
